feat: some renaming

+++ safe_drive/src/msg.rs
@@ -327,7 +327,7 @@ impl<const N: usize> RosString<N> {
         if self.0.data.is_null() {
             &[]
         } else {
-            let s = unsafe { std::slice::from_raw_parts(self.0.data, self.0.size as usize) };
+            let s = unsafe { std::slice::from_raw_parts(self.0.data, self.0.size) };
             s
         }
     }
@@ -336,7 +336,7 @@ impl<const N: usize> RosString<N> {
         if self.0.data.is_null() {
             &mut []
         } else {
-            let s = unsafe { std::slice::from_raw_parts_mut(self.0.data, self.0.size as usize) };
+            let s = unsafe { std::slice::from_raw_parts_mut(self.0.data, self.0.size) };
             s
         }
     }
@@ -409,7 +409,7 @@ impl<const STRLEN: usize, const SEQLEN: usize> RosStringSeq<STRLEN, SEQLEN> {
         if self.0.data.is_null() {
             &[]
         } else {
-            let s = unsafe { std::slice::from_raw_parts(self.0.data, self.0.size as usize) };
+            let s = unsafe { std::slice::from_raw_parts(self.0.data, self.0.size) };
             unsafe { transmute::<&[rosidl_runtime_c__String], &[RosString<STRLEN>]>(s) }
         }
     }
